+++ R/ResamplingCustom.R
@@ -42,8 +42,8 @@ ResamplingCustom = R6Class("ResamplingCustom", inherit = Resampling,
       super$initialize(id = "custom", duplicated_ids = TRUE, man = "mlr3::mlr_resamplings_custom")
     },
 
-    instantiate = function(task, train_sets = NULL, test_sets = NULL) {
-      assert_task(as_task(task))
+    instantiate = function(task, train_sets, test_sets) {
+      task = assert_task(as_task(task))
       assert_list(train_sets, types = "atomicvector", any.missing = FALSE)
       assert_list(test_sets, types = "atomicvector", len = length(train_sets), any.missing = FALSE, null.ok = TRUE)
       self$instance = list(train = train_sets, test = test_sets)