+++ tests/testthat/test_Resampling.R
@@ -1,5 +1,20 @@
 context("Resampling")
 
+test_that("re-instantiating", {
+  t1 = tsk("iris")
+  t2 = tsk("boston_housing")
+  r = rsmp("cv", folds = 2)
+
+  expect_resampling(r$instantiate(t1), task = t1)
+  expect_resampling(r$instantiate(t2), task = t2)
+
+  r = rsmp("custom")
+  expect_error(r$instantiate(t1), "missing")
+
+  expect_resampling(r$instantiate(t1, train_sets = list(1), test_sets = list(1)), task = t1)
+  expect_resampling(r$instantiate(t2, train_sets = list(1), test_sets = list(2)), task = t2)
+})
+
 test_that("param_vals", {
   task = tsk("iris")
   r = rsmp("bootstrap", repeats = 100L, ratio = 1)
@@ -40,4 +55,25 @@ test_that("hashing", {
     with_seed(124L, r$instantiate(task))
     expect_false(identical(r$hash, hash))
   }
+})
+
+test_that("integer grouping col (#396)", {
+  df = data.frame(
+    id = rep(1L:10L, each = 2),
+    x = rnorm(20)
+  )
+
+  tsk = TaskRegr$new(id = "task", backend = df, target = "x")
+  tsk$set_col_role("id", "group")
+
+  bs = rsmp("bootstrap", repeats = 10L, ratio = 1)
+  bs$instantiate(tsk)
+
+  set = bs$train_set(1)
+  expect_integer(set)
+  expect_true(all(map_lgl(split(seq_row(df), f = df$id), function(x) all(x %in% set) || all(x %nin% set))))
+
+  set = bs$test_set(1)
+  expect_integer(set)
+  expect_true(all(map_lgl(split(seq_row(df), f = df$id), function(x) all(x %in% set) || all(x %nin% set))))
 })