+++ tests/testthat/test_mlr_resampling_bootstrap.R
@@ -6,7 +6,7 @@ test_that("bootstrap has duplicated ids", {
 })
 
 test_that("stratification", {
-  data = data.table(y = rep(letters[1:2], times = c(90, 10)), x1 = runif(100), x2 = rep(LETTERS[1:2], times = c(50, 50)))
+  data = data.table(y = factor(rep(letters[1:2], times = c(90, 10))), x1 = runif(100), x2 = rep(LETTERS[1:2], times = c(50, 50)))
   b = as_data_backend(data)
   task = TaskClassif$new("stratify_data", b, target = "y")
   task$col_roles$stratum = task$target_names