+++ tests/testthat/test_mlr_resampling_repeated_cv.R
@@ -17,7 +17,7 @@ test_that("folds first, then repetitions", {
 })
 
 test_that("stratification", {
-  data = data.table(y = rep(letters[1:2], times = c(90, 10)), x1 = runif(100), x2 = rep(LETTERS[1:2], times = c(50, 50)))
+  data = data.table(y = factor(rep(letters[1:2], times = c(90, 10))), x1 = runif(100), x2 = rep(LETTERS[1:2], times = c(50, 50)))
   task = TaskClassif$new("stratify_data", data, target = "y")
   task$col_roles$stratum = task$target_names