+++ NEWS.md
@@ -1,3 +1,16 @@
+# mlr3 0.1.5
+
+* Switched from package `Metrics` to package `mlr3measures` for the
+  implementation of performance measures.
+* Measures can now calculate all scores using micro or macro averaging (#400).
+* Measures can now return a customizable performance score instead of `NA` if
+  they are not defined for a certain input.
+* Character columns are now treated differently from factor columns.
+  In the long term, these are supposed to store text.
+* Fixed a bug triggered by integer grouping variables in `Task` (#396).
+* `benchmark_grid()` now accepts instantiated resamplings under certain
+  conditions.
+
 # mlr3 0.1.4
 
 * `Task$set_col_roles()` and `Task$set_row_roles()` are now deprecated.